+++ packages/plugin-ext/src/plugin/plugin-context.ts
@@ -174,6 +174,8 @@ import { ThemingExtImpl } from './theming';
 import { CommentsExtImpl } from './comments';
 import { CustomEditorsExtImpl } from './custom-editors';
 import { WebviewViewsExtImpl } from './webview-views';
+import { PluginPackage } from '../common';
+import { Endpoint } from '@theia/core/lib/browser/endpoint';
 
 export function createAPIFactory(
     rpc: RPCProtocol,
@@ -1075,9 +1077,15 @@ export class Plugin<T> implements theia.Plugin<T> {
     constructor(protected readonly pluginManager: PluginManager, plugin: InternalPlugin) {
         this.id = plugin.model.id;
         this.pluginPath = plugin.pluginFolder;
-        this.pluginUri = URI.parse(plugin.pluginUri);
         this.packageJSON = plugin.rawModel;
         this.pluginType = plugin.model.entryPoint.frontend ? 'frontend' : 'backend';
+
+        if (this.pluginType === 'frontend') {
+            const { origin } = new Endpoint();
+            this.pluginUri = URI.parse(origin + '/' + PluginPackage.toPluginUrl(plugin.model, ''));
+        } else {
+            this.pluginUri = URI.parse(plugin.pluginUri);
+        }
     }
 
     get isActive(): boolean {